Move window picker exclusive API files back to picker module
