+++ src/window_picker/mod.rs
@@ -1,9 +1,9 @@
 mod input_capture;
+mod native_hooks;
+mod native_ui;
 #[cfg(test)]
 mod tests;
 #[allow(clippy::module_inception)]
 mod window_picker;
 
-pub(crate) use input_capture::{G_INPUT_CAPTURE, InputDisposition, InputEvent, NativeInputSession};
 pub use window_picker::WindowPicker;
-pub(crate) use window_picker::{PickerSessionUi, PickerUi, SelectionDialogChoice, selection_dialog_content};
